requirements.txt
numpy==1.19.2
urllib3==1.25.11
scikit_image==0.17.2
tensorflow==1.15.0
keras_nightly==2.5.0.dev2021032900
pandas==1.1.3
keras==2.4.3
Pillow==8.2.0
skimage==0.0
